Iterating on fues.
